Flag Validation
Detailed Description:
Prior to parsing the flag validation procedure is run. All flags mentioned in the .dgl file are extracted and then the prog searches for the UDialogueFlag asset with the same name. Is the asset is not found, the flag name is added to the list of missing flags.