Added a data table for efficient flag lookup
Detailed Description:
During the runtime, when CondTrees are evaluated, the flags are wrapped in conditional tokens and contains their names as referenced in the correponding .dgl file (e.g. LikesPizza). The corresponding data asset in the editor might have a different name (e.g. DF_LikesPizza_Route1). The data table maps the name in .dgl to the asset name.