Runtime Debug Sustem
Detailed Description:
Added the backbone of the runtime debug system. It is supposed to test DebugManager and DialogueAssets. Is contains two widgets and a DebugController. Debug Controller is spawned by the DialogueUIManager and owns the widgets. DialogueManager sends the UI manager a render request, and the UI manager then redirects the rendering part to the DebugController.