Tree Wrapped for Cond Trees
Detailed Description:
The tree wrapper UObject holds an FString that represents the conditional expression and a unique ptr to the recursive tree structure. Unique ptrs are not serialisable, so upon closing the editor after parsing, the tree is not saved with the dialogue asset. For that every time the tree is needed is has to be re-parsed if it's not in the cache already. For that, every time the runtime attempts to evaluate the tree, the root node it checked and the tree is re-parsed if it's nullptr. The benchmarks for the parsing at runtime are yet to be evaluated. But considering that the trees are going to be quite shallow and few and far between, re-parsing should not be much of a bottleneck. Alternative approach: model the tree with UObject nodes. This creates memory overhead and only makes sense for the in-editor integration with visual graphs and such.